Assistant checkpoint: Add AMC problems mode with random problem selection
Assistant generated file changes:
- client/src/pages/amc.tsx: Create AMC problems page
- server/routes.ts: Add random AMC8 problem endpoint
- client/src/App.tsx: Add AMC route

---

User prompt:

I would like to add a brand new mode to the app, it needs to be on a different page or different "tab", since it will require most of the screen to display the question.

Once that page is loaded, we want to show two buttons: AMC 8 and AMC 10. For now, leave AMC 10 disabled. If user click on AMC 8, it would randomly select 1 problem from the problems' table and display a page section. This page section is like a "div", the content of the "div" will be coming from the question_html field.